Small changes to 2 methods + 2 new examples
Changed 2 `write()` methods, first parameter (address) changed from `int` to `uint16_t` to be sure that it will be 16 bits wide.

The 2 new examples explore the storage & recall (both from software & from power changes) of the chip.

+++ src/EERAM_47XXX.cpp
@@ -102,7 +102,7 @@ void EERAM_47XXX::memoryRecall() {
   writeConfigByte(REG_CONFIG_ADDRESS::COMMAND, COMMAND_EXEC::RECALL);
 }
 
-MEMORY_OP_STATUS EERAM_47XXX::write(int addressPointer, uint8_t singleByte) {
+MEMORY_OP_STATUS EERAM_47XXX::write(uint16_t addressPointer, uint8_t singleByte) {
   if(addressPointer < 0) {return(MEMORY_OP_STATUS::ADDRESS_DNE);}
   if(addressPointer >= m_chipSize) {return(MEMORY_OP_STATUS::ADDRESS_OVER);}
   Wire.beginTransmission(m_address);
@@ -113,7 +113,7 @@ MEMORY_OP_STATUS EERAM_47XXX::write(int addressPointer, uint8_t singleByte) {
   return(MEMORY_OP_STATUS::WRITE_THEORY_OK);
 }
 
-MEMORY_OP_STATUS EERAM_47XXX::write(int addressPointer, uint8_t* dataArray, size_t arraySize, bool overwrite) {
+MEMORY_OP_STATUS EERAM_47XXX::write(uint16_t addressPointer, uint8_t* dataArray, size_t arraySize, bool overwrite) {
   if(addressPointer < 0) {return(MEMORY_OP_STATUS::ADDRESS_DNE);}
   if(addressPointer >= m_chipSize) {return(MEMORY_OP_STATUS::ADDRESS_OVER);}
   if(!overwrite && (addressPointer + arraySize) >= m_chipSize) {return(MEMORY_OP_STATUS::ADDRESS_ROLLOVER);}
